convert the demo_app.js to an MVC architecture
