+++ src/app.ts
@@ -1,0 +1,19 @@
+import express from 'express';
+import bodyParser from 'body-parser';
+import path from 'path';
+import indexRouter from './routes/index';
+
+const app = express();
+const port = 8080;
+
+app.set('view engine', 'ejs');
+app.set('views', path.join(__dirname, 'views'));
+
+app.use(bodyParser.urlencoded({ extended: true }));
+app.use(express.static(path.join(__dirname, 'public')));
+
+app.use('/', indexRouter);
+
+app.listen(port, () => {
+  console.log(`Server running at http://localhost:${port}`);
+});
