--- src/demo_app.ts
@@ -1,187 +1,0 @@
-import { Ollama } from 'ollama';
-import express, { Request, Response } from 'express';
-import bodyParser from 'body-parser';
-import path from 'path';
-import { fileURLToPath } from 'url';
-import { dirname } from 'path';
-import fs from 'fs';
-
-console.log('Starting DON Chat demo app...');
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-
-console.log('dirname:', __dirname);
-
-// Read settings from settings.json
-const settingsPath = path.join(__dirname, '../settings.json');
-const settings = JSON.parse(fs.readFileSync(settingsPath, 'utf-8'));
-
-const app = express();
-const port = 8080;
-
-interface ChatEntry {
-  role: 'user' | 'assistant';
-  content: string;
-}
-
-let chatHistory: ChatEntry[] = [];
-
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(express.static(path.join(__dirname, 'public')));
-
-app.get('/', (req: Request, res: Response) => {
-  const chatHtml = chatHistory.map(entry => `
-    <p><strong>${entry.role === 'user' ? 'You' : 'DON'}:</strong> ${entry.content}</p>
-  `).join('');
-
-  res.send(`
-    <!DOCTYPE html>
-    <html lang="en">
-    <head>
-      <meta charset="UTF-8">
-      <meta name="viewport" content="width=device-width, initial-scale=1.0">
-      <title>DON Chat</title>
-      <link rel="stylesheet" href="${settings.stylePath}">
-    </head>
-    <body>
-      <div class="title-bar">
-        DON Chat
-        <a href="/settings" class="settings-icon">&#9881;</a>
-      </div>
-      <div class="chat-container" id="chat-container">
-        ${chatHtml}
-      </div>
-      <div class="input-container">
-        <form action="/ask" method="post" onsubmit="showLoading()">
-          <input type="text" id="question" name="question" placeholder="Type your question here...">
-          <button type="submit">Ask</button>
-        </form>
-        <form action="/clear" method="post" style="margin-left: 10px;">
-          <button type="submit">Clear Chat</button>
-        </form>
-      </div>
-      <div id="loading">⏳ Loading...</div>
-      <script>
-        function showLoading() {
-          document.getElementById('loading').style.display = 'block';
-        }
-        function scrollToBottom() {
-          const chatContainer = document.getElementById('chat-container');
-          chatContainer.scrollTop = chatContainer.scrollHeight;
-        }
-        window.onload = scrollToBottom;
-      </script>
-    </body>
-    </html>
-  `);
-});
-
-app.get('/settings', (req: Request, res: Response) => {
-  const lightSelected = settings.stylePath === '/styles.css' ? 'selected' : '';
-  const darkSelected = settings.stylePath === '/styles-dark.css' ? 'selected' : '';
-
-  res.send(`
-    <!DOCTYPE html>
-    <html lang="en">
-    <head>
-      <meta charset="UTF-8">
-      <meta name="viewport" content="width=device-width, initial-scale=1.0">
-      <title>Settings</title>
-      <link rel="stylesheet" href="${settings.stylePath}">
-    </head>
-    <body>
-      <div class="title-bar">Settings</div>
-      <form action="/settings" method="post">
-        <label for="style">Select Style:</label>
-        <select id="style" name="style">
-          <option value="/styles.css" ${lightSelected}>Light</option>
-          <option value="/styles-dark.css" ${darkSelected}>Dark</option>
-        </select>
-        <button type="submit">Save</button>
-      </form>
-    </body>
-    </html>
-  `);
-});
-
-app.post('/settings', (req: Request, res: Response) => {
-  const newStylePath = req.body.style;
-  settings.stylePath = newStylePath;
-  fs.writeFileSync(settingsPath, JSON.stringify(settings, null, 2));
-  res.redirect('/');
-});
-
-app.post('/ask', async (req: Request, res: Response) => {
-  const question = req.body.question;
-  chatHistory.push({ role: 'user', content: question });
-
-  try {
-    const ollama = new Ollama({ host: settings.ollamaHost });
-    const response = await ollama.chat({
-      model: settings.model,
-      messages: chatHistory,
-    });
-    const answer = response.message.content;
-    chatHistory.push({ role: 'assistant', content: answer });
-
-    const chatHtml = chatHistory.map(entry => `
-      <p><strong>${entry.role === 'user' ? 'You' : 'DON'}:</strong> ${entry.content}</p>
-    `).join('');
-
-    res.send(`
-      <!DOCTYPE html>
-      <html lang="en">
-      <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>DON Chat</title>
-        <link rel="stylesheet" href="${settings.stylePath}">
-      </head>
-      <body>
-        <div class="title-bar">
-          DON Chat
-          <a href="/settings" class="settings-icon">&#9881;</a>
-        </div>
-        <div class="chat-container" id="chat-container">
-          ${chatHtml}
-        </div>
-        <div class="input-container">
-          <form action="/ask" method="post" onsubmit="showLoading()">
-            <input type="text" id="question" name="question" placeholder="Type your question here...">
-            <button type="submit">Ask</button>
-          </form>
-          <form action="/clear" method="post" style="margin-left: 10px;">
-            <button type="submit">Clear Chat</button>
-          </form>
-        </div>
-        <div id="loading">⏳ Loading...</div>
-        <script>
-          function showLoading() {
-            document.getElementById('loading').style.display = 'block';
-          }
-          function scrollToBottom() {
-            const chatContainer = document.getElementById('chat-container');
-            chatContainer.scrollTop = chatContainer.scrollHeight;
-          }
-          window.onload = scrollToBottom;
-        </script>
-      </body>
-      </html>
-    `);
-  } catch (error) {
-    res.send(`
-      <p>Error communicating with Ollama: ${error}</p>
-      <a href="/">Try again</a>
-    `);
-  }
-});
-
-app.post('/clear', (req: Request, res: Response) => {
-  chatHistory = [];
-  res.redirect('/');
-});
-
-app.listen(port, () => {
-  console.log(`Server running at http://localhost:${port}`);
-});
