+++ src/models/chatEntry.ts
@@ -1,0 +1,4 @@
+export interface ChatEntry {
+    role: 'user' | 'assistant';
+    content: string;
+}
