+++ src/views/settings.ejs
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Settings</title>
+  <link rel="stylesheet" href="<%= stylePath %>">
+</head>
+<body>
+  <div class="title-bar">Settings</div>
+  <form action="/settings" method="post">
+    <label for="style">Select Style:</label>
+    <select id="style" name="style">
+      <option value="/styles.css" <%= lightSelected %>>Light</option>
+      <option value="/styles-dark.css" <%= darkSelected %>>Dark</option>
+    </select>
+    <button type="submit">Save</button>
+  </form>
+</body>
+</html>
