+++ tsconfig.json
@@ -1,12 +1,12 @@
 {
     "compilerOptions": {
       "target": "ES6",
-      "module": "ES2022",
+      "module": "CommonJS",
       "moduleResolution": "node",
       "strict": true,
       "esModuleInterop": true,
       "skipLibCheck": true,
       "outDir": "./dist"
     },
-    "include": ["src/**/*"]
+    "include": ["src/**/*", "app.ts"]
   }
